build_p4_program.nix: add execName, kernelModule, remove symlinks
The new option execName allows building the program under a name
that differs from p4Name. This allows adding variants of the same
P4 program to an environment without causing conflicts.

For the same reason, the symlinks to commands in the SDE package are
removed.

The new option kernelModule loads the requested kernel module before
starting bf_switchd.