Refactor kernel module support
Compilation of kernel modules is now detached from the SDE package.
This avoids re-building of the entire SDE if a kernel is added or
modified.  The modules for each kernel are stored in separate
packages.  This allows installing only the modules required by a
particular system.

The buildP4Program support function now (again) only runs bf_switchd,
without loading or checking modules.  Instead, the P4 program
derivation now has a support function which creates a shell-wrapper
around the main executable which loads the required module for the
local kernel.  This function must be called while on the system on
which the P4 program will be run.