Eliminate p4studio from build process
This is a major re-work of the build process to no longer use the
p4studio builder.  The former bf-sde package is broken up into
separate packages for each component of the SDE

   * bf-syslibs
   * bf-utils
   * bf-drivers
   * bf-diags (currently not actually used)
   * bf-platforms
   * p4c (compiler binaries)
   * tofino-model (model binary)
   * tools (run_* and p4_build scripts)

The former bf-sde package is implemented as a user environment which
combines the relevant packages for the components that expect certain
files to be present in a tree rooted at SDE/SDE_INSTALL.

All semantics of the bf-sde package have been preserved (e.g. the
buildP4Program attribute).