Revert c58512
Add the symlinks for the tofino and tofino_pd_api in bf-drivers
again. As long as this doesn't create any naming conflicts, it should
be safe. Scripts that add those directories to sys.path will still
work as expected. A script written with the assumption that the
directories don't need to be added explicitely need to be patched or
wrapped to set PYTHONPATH.