bf-drivers: remove "dev" output for runtime version
The header files in "dev" must not be exposed in the runtime version
due to the NDA.  The runtime SDE doesn't depend directly to "dev" but
indirectly through nix-support.  We simply don't create a "dev" output
for runtime and delete the header files from "out" (like the other
forbidden stuff).