Reorganize build of kernel modules
Module packagesare now in pkgs.kernel-modules as a set with the same
attributes as the kernel definitions in kernels/default.nix.

Most of the module support functions are no longer needed. Only retain
"modulesForLocalKernels", which is delegated to
kernels/local-kernel.nix.

The wrapper derivation exposes the modules package used to build it as
the "modules" passthru attribute.