Create a more robust SDE environment
The full and runtime versions are now created from a single function
bf-sde/sde/default.nix.  The "tools" sub-package is built implicitly
by that function and is no longer part of the "pkgs" attribute. The
tools that it provides are wrapped to set SDE/SDE_INSTALL, i.e. it is
no longer necessary to do that explicitely before calling them. The
wrappers also set PATH to make the tools self-contained.

The  buildP4Program, moduleWrapper  and runTest  helper functions  are
simplified to make use of the new tools.

The veth* scripts are now taken from ptf-utils.

Use a new approach for the sudo problem. Instead of the mock-sudo
hack, replace sudo with /usr/bin/sudo in the relevant scripts. This
keeps everything else pure. That will have to be changed should we
ever build for NixOS.