Add multi BSP support
Up to now, the bf-platforms sub-package only supported the reference
BSP. This commit adds support for additional BSPs.  The single "bsp"
attribute in the SDE version set is replaced by a set "bsps", with one
attribute per BSP. Each BSP produces one or more "baseboards",
i.e. instances of the bf-platforms sub-package.  Each baseboard
provides support for one or more platforms. Ther is a unique mapping
of a platform to a baseboard to a BSP.

The runtimeEnv sub-package is replaced by a support function which
creates a baseboard-dependent runtime environment.  The buildP4Program
function also takes a new argument "platform" to specify which
baseboard needs to be included (via the mapping function).

The "env" Makefile target tries to infer the platform from
/etc/machine.conf if it exists.