Add support for the SAL on the aps_bf2556 baseboard
The Switch Abstraction Layer (SAL) supports the Marvell gear box on
the APS BF2556X_1T.  It is part of the bf-platforms package derived
from the APS BSP and consists of one binary (salRefApp) and a set of
Python modules derived from a protobuf definition for the
communication with the salRefApp.  On that platform, bf_switchd is not
started as a standalone process. Instead, it is launched by salRefApp
in order to be able to program the Tofino ports that are connected to
the gear box (ports 1/0 through 16/0).  The buildP4Program function
supports this by generating a separate wrapper script for the
stordis_bf2556x_1t platform.