Use CMake build system for 9.6.0
This commit switches to the new CMake-based build procedure for SDE
9.6.0. The overall structure of the Nix packages has been retained but
it requires more code to keep the modular structure as with the
autotools-based build mechanism.  Support functions for the CMake
build have been added in bf-sde/build-system, see the comments in
default.nix for details.

9.6.0 is reinstated as bf-sde.latest.

The old p4_build.sh has been kept for the time being. It uses the
p4-build component, which has not been ported to CMake by Intel and is
still part of 9.6.0.  There is a CMake-based replacement of
p4_build.sh, which is used by p4studio to build the example programs,
but it is not clear whether this is how regular P4 programs should be
built in the future.

Care has been taken to not change any derivations for previous SDE
versions to avoid regressions at the price of some bloat in Nix
expressions.  Consider cleaning up the code in future versions.