Add support for SDE 9.7.0
This release removes all artifacts of the old autotools-based build
system, including the environment to compile P4 programs.  This is
replaced with a CMake-based system as well, which is accounted for by
using a new p4_build.sh script.

Up to 9.6.0, bf-utils still included the installation procedure for
the bf_rt_python modules. That was somewhat bogus because the modules
themselves have been moved to the bf-drivers package in 9.3.0.  The
Nix packaging requires those modules to be part of the bf-utils
package. From 9.3.0 up to 9.6.0 it was enough to copy those modules
from bf-drivers when building bf-utils.  With 9.7.0, we also have to
force the installation of the copied modules. This entire problem
exists, because the monolithic p4studio install procedure assumes that
everything is installed in a single location $SDE_INSTALL, which is
not the case for the Nix packages.

A few additional tweaks are necessary to compensate for various
changes in the CMake procedures of some components.