Add envCommand and envStandalone to replace "make env"
The envCommand support function provides a user-friendly replacement
of the "make env" mechanism. It creates a command that starts an SDE
shell and allows easier overrides of packages and python modules than
using the "inputFn" argument to mkShell directly.

The envStandalone function builds an installer for envCommand that
does not require network access at install time.