Add initial multi-target support
The buildP4Program support function accepts an additional argument
"target" to select the P4 target to build for (tofino or tofino2).

The architecture is derived from the target where needed (e.g. for the
tofino-model or bf_switchd) since there currently is a 1-to-1
correspondence between the two.  We might have to re-visit this.