buildP4Program: remove unwanted dependencies of P4 artifacts
Starting with the cmake-based builds in 9.7.0, the package for the
compiler artifacts of a P4 program contains two additional JSON files
source.json and fronted-ir.json that contain references to the P4
source code as well as the P4 compiler package. The former is a
nuisance but the latter is a problem because the compiler is not
allowed to be packaged for a binary distribution.  It appears that
these files are not needed at runtime (and, in fact, also exist in
older SDE versions but are not installed by the original p4_build.sh
script). They are omitted from the artifacts package by default but
can be retained by setting the new argument pureArtifacts to the
buildP4Program function to false.