release-manager: Remove kernel dependency for Tofino model operation
Use a feature of the handling of unknown kernels to let the
release-manager install the model platform on any kernel.

The module unloading code is removed from the model-specific module
wrapper because that is already taken care of by the script
run-model.sh hat starts bf_switchd on the model.

Also fix impurities in run-model.sh.