Overhaul CMake build support
The new approach enables standalone builds of the SDE components by
adding some elements of the original top-level CMake machinery to the
components in a less convoluted way. Instead of overloading the
preConfigure phase, the modifications are performed by transforming
the source archives and then using those modified sources as input to
the derivations.

More use of CMake primitives to apply additional changes to make the
standalone builds to work allows to remove a lot of patches, which
should make it easier to integrate future SDE versions.