Add support for the Newport Tofino2 baseboard
bf-sde/bf-platforms/reference.nix now creates an additional baseboard
called "newport" that supports the Tofino2 reference platfrom
AS9516-32D.

The main complication with this baseboard is that it requires building
a kernel module to drive a FPGA-based I2C controller. We use a new
mechanism that should be able to collect kernel modules from arbitrary
components togehter with the standard modules from the bf-drivers
component, see the comment for the "drvsWithKernelModules" attribute
in bf-sde/default.nix.