Support custom portinfo file on model platform
A custom portinfo file can be passed to the command generated by
buildP4Program() on the model platform through the
TOFINO_MODEL_PORTINFO environment variable.  If unset, the script
behaves as in previous versions, i.e. no "-f" option is passed to the
tofino_model binary and the standard set of veth pairs is created by
the veth_setup.sh script.  If a custom portinfo file is used, the
veth_{setup,teardown}.sh scripts generate only the veth pairs
requested by the PortToVeth object in the file.