Add support for the Inventec D10064 for 9.7.x
Inventec has re-branded the D5264 to D10064. This commit adds the BSP
for SDE versions 9.7.x. Inventec is still using ONL as its base system
despite Intel having ceased support for ONL. The BSP depends on the
presence of a set of kernel modules that are not part of the
BSP. These modules perform the enumeration of the I2C bus and provide
access to the QSFP modules as well as the system EEPROM. They are
written for the 4.19 kernel shipped with ONL and require patching to
work on 5.10.

As it turns out, only the access to the EEPROM that contains the
system information is used by the BSP.  The BSP does not implement any
of the APIs related to QSFP or hardware sensors. Thus, these objects
are invisible to the SDE (the "qsfp" context is missing completely
from the bf_pltfm UCLI and the chss_mgmt context does not provide
access to anything not related directly to the Tofino ASIC).

Thus, this BSP behaves very much like a "bsp-less" system from the
perspective of the SDE.  The QSFP modules and hardware sensors can be
accessed in a rudimentary fashion through sysfs objects created by the
ONL modules:

power suplies
  /sys/bus/i2c/devices/0-0055/

QSFP modules
  /sys/devices/virtual/swps/