Add ONL-based kernel mdules for the Inventec D10064
The Inventec BSP depends on the presence of a set of kernel modules
that are not part of the BSP but are maintained on the
Inventec-specific part of ONL. These modules perform the enumeration
of the I2C bus and provide access to the QSFP modules as well as the
system EEPROM. They are written for the 4.19 kernel shipped with ONL
and require patching to work on 5.10.

As it turns out, only the access to the EEPROM that contains the
system information is used by the BSP.  The BSP does not implement any
of the APIs related to QSFP or hardware sensors. Thus, these objects
are invisible to the SDE (the "qsfp" context is missing completely
from the bf_pltfm UCLI and the chss_mgmt context does not provide
access to anything not related directly to the Tofino ASIC).

As a consequence this BSP behaves very much like a "bsp-less" system
from the perspective of the SDE.  The QSFP modules and hardware
sensors can be accessed in a rudimentary fashion through sysfs objects
created by the ONL modules:

power suplies
  /sys/bus/i2c/devices/0-0055/

QSFP modules
  /sys/devices/virtual/swps/

This will most likely be the final state of support of this platform
for SDE 9.7. Inventec is currently not porting the BSP to a newer
version and it appears that they will also not update the 9.7 version
to work without the ONL components.