Add a pseudo-kernel to support the model platforms like regular platforms
The derivation of the pseudo-kernel contains scripts that terminate
with an error when attempting to load a kernel module.  This
derivation is returned by the modulesForKernel() support function when
called with the kernel release "none". This allows to treat the model
platform like regular platforms when a kernel module derivation is
expected to exist.  It is also expected that P4 programs are built
with requiredKernelModule = null for the model such that the
pseudo-kernel derivation is never actually called.