Add support for bsp-less platforms
A platform is declared to be bsp-less by setting baseboard = null in
bf-sde/bf-platforms/properties.nix. In that case, no BSP package is
required to exist and the run-time environment for P4 programs
compiled for the platform does not contain a BSP.  A "board port map"
JSON file is required to exist for such platforms. When bf_switchd is
started with the port map file and no BSP present, the UCLI does not
contain the bf_pltfm context and no platform-specific features are
available. Most transceiver plugins are still expected to work.

This commit contains the Netberg Aurora 710 as a (random) example of a
bsp-less platform.