Add support to build non-standard kernel modules
Each kernel comes with a set of kernel modules defined by the release
managers of the respective Linux distribution. The SDE or a BSP may
require a module that is not in this list. Instead of re-configuring
and re-building the entire kernel to include the missing modules, we
simply build the modules from the kernel source as out-of-tree
modules. Like the drvsWithKernelModules mechanism, the list of modules
can be applied for all baseboards or for a specific baseboard
only. The build of the modules is also triggered in build-modules.nix.

The kernel source is accessed through passthru.source of the buildTree
kernel attribute. This is done to allow a kernel flavor-specific
method to be used to create the source code from
distribution-dependent resources. Currently, only the Debian flavor is
supported where the source code is supplied by a single .deb file.