Fix subtle bug in merging of SDE specs
When merging the common specs with the SDE-specific specs we must be
careful to not descend into derivations since merging derivations
makes no sense and can lead to corruption. In the current code, this
happens when python2 is overriden by python3. One possible effect is
that a module that doesn't exist in python3 inherits the module from
python2.

To fix, use recursiveUpdateUntil instead of recursiveUpdate and add a
check that an attribute set is not a derivation.