Simplify autotools/CMake integration, fix optional vs optionals
In the original move to CMake-based builds care was taken to not
change the derivations for the older autotools-based builds. These
precautions are removed here to make the code (hopefully) less
convoluted.

There were also many instances of an incorrect use of lib.optional
that created nested instead of flat lists. In all of these cases, an
implicit flattening took place, i.e. the changes in this commit are
not semantically relevant.