Move to Nixpkgs 22.05 (099cb1)
Even new SDE versions still require some fairly old versions of
dependencies (e.g. Thrift and protobuf).

The standard compiler in 22.05 is GCC 11, which can cause problems
when linking libraries with objects compiled with older versions, as
is required by most SDEs. The symptom of this issue is that the symbol
std::__throw_bad_array_new_length cannot be resolved. The solution is
to recompile these libraries with GCC 10.

Python2 is still part of 22.05 but it supports only a small number of
modules. Most of the missing modules can be eliminated by resricting
the features of the scapy module. Also, some self-tests that no longer
pass have to be disabled.