Add baseboard blacklist for kernels
Baseboards that require building additional kernel modules can be
incompatible with some kernels.  This can be made explicit by adding
the baseboard to the kernel's baseboardBlacklist, which causes an
assertion failure when trying to build the combined kernel module
derivation for the baseboard. The blacklisted combinations are
excluded when building pkgs.kernel-modules-baseboards to make Hydra
builds succeed.