sde-env: add --ptf-pkgs option
This option creates the environment variable PTF_PATH which contains
the bin directories for all packages passed with the option. PTF_PATH
is then merged by the run_p4_tests.sh shell wrapper with PATH to make
the packages available to the PTF script. This is typically used to
call external programs.