Improve support for BSP-less platforms
BSP-less mode is now also supported in the sde shell. An additional
mode is added where a platform can be supported by a BSP for a subset
of the available SDE versions and in BSP-less mode for all other
versions.  This feature can be used by setting the proper baseboard
name in bf-sde/bf-platforms/properties.nix but adding a port map file
like for BSP-less mode.

The support function baseboardForPlatform is modified to support the
new BSP-less fallback method and raises exceptions if a platform is
not supported at all or not supported by the specific SDE either in
BSP or BSP-less fallback mode.