Move nixpkgs to master
The nixpkgs-master branch is intended to track nixpkgs master to
facilitate transitions to a new nixpkgs version. The master branch
remains pinned to a specific version of nixpkgs. The idea is to move
to the newest nixpkgs release when it becomes available by merging
nixpkgs-master to master.

This initial commit containes the changes required to move from 22.05
to nixpkgs master.  Most changes concern Python2 which will be
disabled by default in 23.05 for the first time. It's still easy to
support it for SDEs older than 9.7 so there is no immediate need to
drop support for those SDEs.

SDEs from 9.7 to 9.10 require Python3 but also an old version of
protobuf which is not compatible with Python 3.10, which is the
Default Python3 since 22.10. Those SDEs have to use a customised
version of Python 3.9. SDE 9.11 and newer can use the newest protobuf
and grpc versions and hence can use the standard Python3.