bf-sde/build-p4: fix artifacts for platforms with BSP-less fallback
For platforms that are configured to fall back to BSP-less mode for
unsupported SDEs (i.e. those that have a non-null baseboard as well as
a portMap attribute in properties.nix), the build artifacts referenced
the port-map file even for supported SDEs. This effectively forced
BSP-less mode for all SDE versions.

The fix is to apply the port-map file only for unsupported SDEs.