bf-diags: adapt to changes in glibc 2.37
linux/if_packet.h was changed to use a flexible array in struct
sockaddr_ll. This requires bf-diags to put instances of sockaddr_ll at
the end of structs. This commit moves struct members around to
accomplish that.

The change in glibc looks a bit weird. It makes it impossible to have
multiple instances of sockaddr_ll in another struct, for instance. But
maybe it's a kernel private struct and bf-diags is wrong to use it :/