mkOnieInstaller: update onie-debian-nix-installer, optionally include SDE shell
Bump the version of onie-debian-nix-installer to a93b72. Apart from
bug-fixes, this version introducess a non-backward compatible change
of the semantics of the postFs{Create,Install}Cmd arguments, which now
take lists of derivations as values instead of single derivations to
make customization easier.

The mkOnieInstaller support function is changed as follows

   * Replace the partialSlice argument by slice

   * The bootstrap profile now must contain a file kernelID.nix that
     evaluates to the unique kernel ID that matches the kernel in the
     profile

   * Add the withSdeEnv argument (default false). When set to true,
     include an instance of the sde-env command to the system's
     default profile to provide an SDE development environment

   * Add the activate argument (default true). If set to false, the
     service is not activated during the installation of the target
     system via the release-manager