Add Debian 13.1 kernel (6.12.43+deb13)
All kernel modules work with minor modifications.

Somtime between kernels 6.1 and 6.12, the Kbuild environment changed
wrt separated build and source directories for out-of-tree
modules. Previously, the common practice was to use Kbuild with

  make -C <path-to-kbuild> -M=<path-to-build-dir> src=<path-to-source-dir>

This no longer works. It seems like using "src" in this manner was
never officially supported. Starting with 6.13, Kbuild has a new
option "MO" to officially support a split source/destination setup,
presumably to make up for the loss of this "accidental" feature in
previous versions. For 6.12, we work around this problem by always
calling Kbuild with

  make -C <path-to-kbuild> -M=<path-to-source-dir>

which generates the build artifacts in the source directory. For CMake
builds, we also have to change the installation step to install from
the source tree rather than from the binary tree.